extract semver into separate class to be referenceable by github strategy
